requirements.txt
fastapi
uvicorn
pillow
opencv-python
python-dotenv
requests
streamlit
google-genai
python-multipart